+++ lib/audioreleaser.rb
@@ -6,3 +6,4 @@ require 'audioreleaser/track'
 require 'audioreleaser/license'
 require 'audioreleaser/tagger'
 require 'audioreleaser/tag'
+require 'audioreleaser/release'
